Move component playground to own folder #114
