+++ tests/components/component-playground/render.js
@@ -3,7 +3,7 @@ var $ = require('jquery'),
     serialize = require('../../../lib/serialize.js'),
     renderComponent = require('../../helpers/render-component.js'),
     ComponentPlayground =
-      require('../../../components/component-playground.jsx');
+      require('../../../component-playground/component-playground.jsx');
 
 describe('ComponentPlayground component', function() {
   var component,
