+++ tests/components/component-playground/state.js
@@ -2,7 +2,7 @@ var $ = require('jquery'),
     Cosmos = require('../../../cosmos.js'),
     renderComponent = require('../../helpers/render-component.js'),
     ComponentPlayground =
-      require('../../../components/component-playground.jsx');
+      require('../../../component-playground/component-playground.jsx');
 
 describe('ComponentPlayground component', function() {
   var component,
